+++ src/dcegm/interfaces/interface.py
@@ -3,6 +3,7 @@
 import jax
 import jax.numpy as jnp
 import pandas as pd
+from jax import numpy as jnp
 
 from dcegm.interfaces.index_functions import (
     get_state_choice_index_per_discrete_states_and_choices,
@@ -289,3 +290,117 @@ def stochastic_transition_vec(state_choice_vec_dict, func, params):
 
     """
     return func(**state_choice_vec_dict, params=params)
+
+
+def choice_values_for_states(
+    value_solved,
+    endog_grid_solved,
+    state_choice_indexes,
+    params,
+    states,
+    model_config,
+    model_funcs,
+):
+    value_grid_states = jnp.take(
+        value_solved,
+        state_choice_indexes,
+        axis=0,
+        mode="fill",
+        fill_value=jnp.nan,
+    )
+    endog_grid_states = jnp.take(
+        endog_grid_solved,
+        state_choice_indexes,
+        axis=0,
+        mode="fill",
+        fill_value=jnp.nan,
+    )
+
+    def wrapper_interp_value_for_choice(
+        state,
+        value_grid_state_choice,
+        endog_grid_state_choice,
+        choice,
+    ):
+        state_choice_vec = {**state, "choice": choice}
+
+        return interpolate_value_for_state_and_choice(
+            value_grid_state_choice=value_grid_state_choice,
+            endog_grid_state_choice=endog_grid_state_choice,
+            state_choice_vec=state_choice_vec,
+            params=params,
+            model_config=model_config,
+            model_funcs=model_funcs,
+        )
+
+    # Read out choice range to loop over
+    choice_range = model_config["choices"]
+
+    choice_values_per_state = jax.vmap(
+        jax.vmap(
+            wrapper_interp_value_for_choice,
+            in_axes=(None, 0, 0, 0),
+        ),
+        in_axes=(0, 0, 0, None),
+    )(
+        states,
+        value_grid_states,
+        endog_grid_states,
+        choice_range,
+    )
+    return choice_values_per_state
+
+
+def choice_policies_for_states(
+    policy_solved,
+    endog_grid_solved,
+    state_choice_indexes,
+    states,
+    model_config,
+):
+    policy_grid_states = jnp.take(
+        policy_solved,
+        state_choice_indexes,
+        axis=0,
+        mode="fill",
+        fill_value=jnp.nan,
+    )
+    endog_grid_states = jnp.take(
+        endog_grid_solved,
+        state_choice_indexes,
+        axis=0,
+        mode="fill",
+        fill_value=jnp.nan,
+    )
+
+    def wrapper_interp_value_for_choice(
+        state,
+        policy_grid_state_choice,
+        endog_grid_state_choice,
+        choice,
+    ):
+        state_choice_vec = {**state, "choice": choice}
+
+        return interpolate_policy_for_state_and_choice(
+            policy_grid_state_choice=policy_grid_state_choice,
+            endog_grid_state_choice=endog_grid_state_choice,
+            state_choice_vec=state_choice_vec,
+            model_config=model_config,
+        )
+
+    # Read out choice range to loop over
+    choice_range = model_config["choices"]
+
+    choice_values_per_state = jax.vmap(
+        jax.vmap(
+            wrapper_interp_value_for_choice,
+            in_axes=(None, 0, 0, 0),
+        ),
+        in_axes=(0, 0, 0, None),
+    )(
+        states,
+        policy_grid_states,
+        endog_grid_states,
+        choice_range,
+    )
+    return choice_values_per_state
